requirements.txt
discord.py
pyautogui
keyboard
win32